5.4. Протеиновая диета [#156313 #200212]
  1. Выполнила задачу "Протеиновая диета".
  2. Подключен мэвен проект.
