+++ src/main/java/ru/job4j/Main.java
@@ -1,7 +1,7 @@
 package ru.job4j;
 
 public class Main {
-    public static void main(String [] args) {
+    public static void main(String[] args) {
         System.out.println("Hello world");
     }
 }
