+++ src/main/java/ru/job4j/calculator/ArgMethod.java
@@ -4,6 +4,7 @@ public class ArgMethod {
     public static void hello(String name) {
         System.out.println("Hello," + name);
     }
+
     public static void main(String[] args) {
         String name = "User";
         ArgMethod.hello(name);
