+++ src/main/java/ru/job4j/calculator/Fit.java
@@ -2,14 +2,16 @@ package ru.job4j.calculator;
 
 public class Fit {
     public static double manWeight(short height) {
-        double rsl = (height-100)*1.15;
+        double rsl = (height - 100) * 1.15;
         return rsl;
     }
+
     public static double womanWeight(short height) {
-        double rsl = (height-110)*1.15;
+        double rsl = (height - 110) * 1.15;
         return rsl;
     }
-    public static void main(String [] args) {
+
+    public static void main(String[] args) {
         short height = 187;
         double man = Fit.manWeight(height);
         System.out.println("Man 187 is " + man);
