+++ src/main/java/ru/job4j/calculator/X2.java
