+++ src/main/java/ru/job4j/condition/AlertDivByZero.java
