+++ src/main/java/ru/job4j/condition/ChessBoard.java
