+++ src/main/java/ru/job4j/condition/Cinema.java
