+++ src/main/java/ru/job4j/condition/DummyBot.java
