+++ src/main/java/ru/job4j/condition/Game.java
@@ -12,6 +12,7 @@ public class Game {
             System.out.println("Start - tetris");
         }
     }
+
     public static void main(String[] args) {
         Game.menu("tanks");
     }
