+++ src/main/java/ru/job4j/condition/Greeting.java
