+++ src/main/java/ru/job4j/condition/Liken.java
