+++ src/main/java/ru/job4j/condition/LogicNot.java
