+++ src/main/java/ru/job4j/condition/Max.java
