+++ src/main/java/ru/job4j/condition/MultiMax.java
