+++ src/main/java/ru/job4j/condition/PointEq.java
