+++ src/main/java/ru/job4j/condition/SqArea.java
@@ -2,9 +2,9 @@ package ru.job4j.condition;
 
 public class SqArea {
     public static double square(int p, int k) {
-        double h = p/(2*k+2);
-        double l =h*k;
-        double rsl = l*h;
+        double h = p / (2 * k + 2);
+        double l = h * k;
+        double rsl = l * h;
         return rsl;
     }
