+++ src/main/java/ru/job4j/condition/SwitchWeek.java
