+++ src/main/java/ru/job4j/condition/ThreeMax.java
