+++ src/main/java/ru/job4j/condition/TrgArea.java
@@ -2,9 +2,9 @@ package ru.job4j.condition;
 
 public class TrgArea {
     public static double area(double a, double b, double c) {
-        double p = (a+b+c)/2;
-        double first = (p-a)*(p-b)*(p-c);
-        double second = p*first;
+        double p = (a + b + c) / 2;
+        double first = (p - a) * (p - b) * (p - c);
+        double second = p * first;
         double rsl = Math.sqrt(second);
         return rsl;
     }
