+++ src/main/java/ru/job4j/condition/Triangle.java
@@ -2,6 +2,6 @@ package ru.job4j.condition;
 
 public class Triangle {
     public static boolean exist(double ab, double ac, double bc) {
-        return ab+ac>bc && ac+bc>ab && bc+ab>ac;
+        return ab + ac > bc && ac + bc > ab && bc + ab > ac;
     }
 }
