+++ src/main/java/ru/job4j/loop/Board.java
