+++ src/main/java/ru/job4j/loop/Counter.java
@@ -8,6 +8,7 @@ public class Counter {
         }
         return sum;
     }
+
     public static int sumByEven(int start, int finish) {
         int sum = 0;
         for (int i = start; i <= finish; i++) {
@@ -17,6 +18,7 @@ public class Counter {
         }
         return sum;
     }
+
     public static void main(String[] args) {
         System.out.println(sum(0, 10));
         System.out.println(sum(3, 8));
