+++ src/main/java/ru/job4j/loop/Factorial.java
