+++ src/main/java/ru/job4j/loop/Fitness.java
