+++ src/main/java/ru/job4j/loop/LoopFor.java
