+++ src/main/java/ru/job4j/math/MathFunction.java
