+++ src/test/java/ru/job4j/MainTest.java
