+++ src/test/java/ru/job4j/calculator/X2Test.java
