+++ src/test/java/ru/job4j/condition/ChessBoardTest.java
