+++ src/test/java/ru/job4j/condition/MaxTest.java
@@ -1,11 +1,12 @@
 package ru.job4j.condition;
 
 import org.junit.Test;
+
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
-
 public class MaxTest {
+
     @Test
     public void whenMax1To2Then2() {
         int result = Max.max(1, 2);
