+++ src/test/java/ru/job4j/condition/MultiMaxTest.java
