+++ src/test/java/ru/job4j/condition/PointTest.java
@@ -14,6 +14,7 @@ public class PointTest {
         double out = Point.distance(x1, y1, x2, y2);
         Assert.assertEquals(expected, out, 0.01);
     }
+
     @Test
     public void when20to60then4() {
         int x1 = 2;
