+++ src/test/java/ru/job4j/condition/SqAreaTest.java
