+++ src/test/java/ru/job4j/condition/ThreeMaxTest.java
@@ -21,7 +21,7 @@ public class ThreeMaxTest {
 
     @Test
     public void thirdMax() {
-        int rsl = ThreeMax.max(1, 5, 100) ;
+        int rsl = ThreeMax.max(1, 5, 100);
         assertThat(rsl, is(100));
     }
