+++ src/test/java/ru/job4j/condition/TriangleTest.java
@@ -1,6 +1,7 @@
 package ru.job4j.condition;
 
 import org.junit.Test;
+
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
 
@@ -10,6 +11,7 @@ public class TriangleTest {
         boolean result = Triangle.exist(2.0, 2.0, 2.0);
         assertThat(result, is(true));
     }
+
     @Test
     public void whenNotExist() {
         boolean result = Triangle.exist(-2.0, 2.0, 2.0);
