+++ src/test/java/ru/job4j/converter/ConverterTest.java
@@ -16,8 +16,9 @@ public class ConverterTest {
     @Test
     public void whenConvert180RblThen3Dlr() {
         int in = 180;
-        float expected = 3;
+        float expected = 3.0f;
         float out = Converter.rubleToDollar(in);
-        Assert.assertEquals(expected, out);
+        System.out.println(out);
+        Assert.assertEquals(expected, out, 0.1);
     }
 }
