+++ src/test/java/ru/job4j/loop/CounterTest.java
