+++ src/test/java/ru/job4j/loop/FactorialTest.java
