+++ src/test/java/ru/job4j/loop/FitnessTest.java
